+++ src/main/java/com/ManticCafe/potato_core/common/item/weapon/the_last_sword.java
@@ -12,10 +12,11 @@ import net.minecraft.world.level.Level;
 import com.ManticCafe.potato_core.common.entity.entities.DiamondProjectileEntity;
 import com.ManticCafe.potato_core.common.entity.entityhandler;
 import net.minecraft.world.phys.Vec3;
+import com.ManticCafe.potato_core.common.config.configReader;
 
 public class the_last_sword extends SwordItem {
 
-    private static final float PROJECTILE_DAMAGE = 12.0F;
+    private static final int PROJECTILE_DAMAGE = 12;
 
     public the_last_sword() {
         super(infinite_tier.INFINITE_TIER,0,0,new Item.Properties().durability(-1));
@@ -48,7 +49,7 @@ public class the_last_sword extends SwordItem {
             level.addFreshEntity(diamondProjectile);
 
             //技能CD
-            player.getCooldowns().addCooldown(this, 80);
+            player.getCooldowns().addCooldown(this,80);
         }
 
         return InteractionResultHolder.sidedSuccess(itemstack, level.isClientSide());
