+++ src/main/java/com/ManticCafe/potato_core/common/item/weapon/the_last_sword_level8.java
@@ -15,7 +15,7 @@ import net.minecraft.world.phys.Vec3;
 
 public class the_last_sword_level8 extends SwordItem {
 
-    private static final float PROJECTILE_DAMAGE = 64.0F;
+    private static final int PROJECTILE_DAMAGE = 64;
 
     public the_last_sword_level8() {
         super(infinite_tier.INFINITE_TIER,104,4.5F,new Item.Properties().durability(-1));
